test(core): keep the Wave 4 pack tests inside their owning pack
`lint-pack-test-boundary.py` (pack-tests-stay-in-pack) rejected 23
above-pack reaches in the new Wave 4 suites. Repository-level coverage
now sits in the roster and pack-local coverage anchors at packs/core:

- Move test_durable_outputs_and_release.py wholesale to
  tests/roster/test_wave4_durable_outputs_and_release.py; every surface
  it reads (docs/, guides/, the spec pair, the changelog) is
  repository-level.
- Move the projected-vs-agentbundle file_safety byte-parity check and
  the spec/index/plan/workspace lifecycle check into the existing
  close-work roster file.
- Re-anchor test_close_work.py and test_pause_receipts_and_initiative.py
  at the pack root with literal, statically resolvable subject paths.
- Load the contract-amendment subjects from named constants; a path
  built from a function parameter cannot be resolved by the lint.

Spec: docs/specs/close-work-extraction-and-immediate-disposition/spec.md
